+++ CHANGELOG.md
@@ -7,8 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [5.9.2] - 2026-01-11
+
 ### Fixed
 
+* Added -5 penalty checkbox toggle for subsisting after 8 hours of exploration
 * Reset all camping time trackers after advancing 24 hours or more and not just the one from the active party
 
 ## [5.9.1] - 2026-01-11